+++ transport_analysis/analysis/velocityautocorr.py
@@ -5,111 +5,203 @@ VelocityAutocorr --- :mod:`transport_analysis.analysis.VelocityAutocorr`
 This module contains the :class:`VelocityAutocorr` class.
 
 """
-from typing import Union, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from MDAnalysis.analysis.base import AnalysisBase
+from MDAnalysis.core.groups import UpdatingAtomGroup
+from MDAnalysis.exceptions import NoDataError
 import numpy as np
+from transport_analysis.analysis.due import due, Doi
 
 if TYPE_CHECKING:
-    from MDAnalysis.core.universe import Universe, AtomGroup
+    from MDAnalysis.core.universe import AtomGroup
+
+due.cite(
+    Doi("10.21105/joss.00877"),
+    description="Autocorrelation with tidynamics",
+    path="transport_analysis.analysis.velocityautocorr",
+    cite_module=True,
+)
 
 
 class VelocityAutocorr(AnalysisBase):
-    """VelocityAutocorr class.
-
-    This class is used to perform analysis on a trajectory.
+    """
+    Class to calculate a velocity autocorrelation function (VACF).
 
     Parameters
     ----------
-    universe_or_atomgroup: :class:`~MDAnalysis.core.universe.Universe` or :class:`~MDAnalysis.core.groups.AtomGroup`
-        Universe or group of atoms to apply this analysis to.
-        If a trajectory is associated with the atoms,
-        then the computation iterates over the trajectory.
-    select: str
-        Selection string for atoms to extract from the input Universe or
-        AtomGroup
+    atomgroup : AtomGroup
+        An MDAnalysis :class:`~MDAnalysis.core.groups.AtomGroup`.
+        Note that :class:`UpdatingAtomGroup` instances are not accepted.
+    dim_type : {'xyz', 'xy', 'yz', 'xz', 'x', 'y', 'z'}
+        Desired dimensions to be included in the VACF. Defaults to 'xyz'.
+    fft : bool
+        If ``True``, uses a fast FFT based algorithm for computation of
+        the VACF. Otherwise, use the simple "windowed" algorithm.
+        The tidynamics package is required for `fft=True`.
+        Defaults to ``True``.
 
     Attributes
     ----------
-    universe: :class:`~MDAnalysis.core.universe.Universe`
-        The universe to which this analysis is applied
-    atomgroup: :class:`~MDAnalysis.core.groups.AtomGroup`
+    atomgroup : :class:`~MDAnalysis.core.groups.AtomGroup`
         The atoms to which this analysis is applied
-    results: :class:`~MDAnalysis.analysis.base.Results`
-        results of calculation are stored here, after calling
-        :meth:`VelocityAutocorr.run`
-    start: Optional[int]
-        The first frame of the trajectory used to compute the analysis
-    stop: Optional[int]
-        The frame to stop at for the analysis
-    step: Optional[int]
-        Number of frames to skip between each analyzed frame
-    n_frames: int
-        Number of frames analysed in the trajectory
-    times: numpy.ndarray
-        array of Timestep times. Only exists after calling
-        :meth:`VelocityAutocorr.run`
-    frames: numpy.ndarray
-        array of Timestep frame indices. Only exists after calling
-        :meth:`VelocityAutocorr.run`
+    dim_fac : int
+        Dimensionality :math:`d` of the VACF.
+    results.timeseries : :class:`numpy.ndarray`
+        The averaged VACF over all the particles with respect to lag-time.
+        Obtained after calling :meth:`VelocityAutocorr.run`
+    results.vacf_by_particle : :class:`numpy.ndarray`
+        The VACF of each individual particle with respect to lag-time.
+    start : Optional[int]
+        The first frame of the trajectory used to compute the analysis.
+    stop : Optional[int]
+        The frame to stop at for the analysis.
+    step : Optional[int]
+        Number of frames to skip between each analyzed frame.
+    n_frames : int
+        Number of frames analysed in the trajectory.
+    n_particles : int
+        Number of particles VACF was calculated over.
     """
 
     def __init__(
         self,
-        universe_or_atomgroup: Union["Universe", "AtomGroup"],
-        select: str = "all",
-        # TODO: add your own parameters here
-        **kwargs
-    ):
+        atomgroup: "AtomGroup",
+        dim_type: str = "xyz",
+        fft: bool = True,
+        **kwargs,
+    ) -> None:
         # the below line must be kept to initialize the AnalysisBase class!
-        super().__init__(universe_or_atomgroup.trajectory, **kwargs)
+        super().__init__(atomgroup.universe.trajectory, **kwargs)
         # after this you will be able to access `self.results`
         # `self.results` is a dictionary-like object
         # that can should used to store and retrieve results
         # See more at the MDAnalysis documentation:
         # https://docs.mdanalysis.org/stable/documentation_pages/analysis/base.html?highlight=results#MDAnalysis.analysis.base.Results
 
-        self.universe = universe_or_atomgroup.universe
-        self.atomgroup = universe_or_atomgroup.select_atoms(select)
+        if isinstance(atomgroup, UpdatingAtomGroup):
+            raise TypeError(
+                "UpdatingAtomGroups are not valid for VACF computation"
+            )
+
+        # args
+        self.dim_type = dim_type
+        self._parse_dim_type()
+        self.fft = fft
+
+        # local
+        self.atomgroup = atomgroup
+        self.n_particles = len(self.atomgroup)
 
     def _prepare(self):
-        """Set things up before the analysis loop begins"""
-        # This is an optional method that runs before
-        # _single_frame loops over the trajectory.
-        # It is useful for setting up results arrays
-        # For example, below we create an array to store
-        # the number of atoms with negative coordinates
-        # in each frame.
-        self.results.is_negative = np.zeros(
-            (self.n_frames, self.atomgroup.n_atoms),
-            dtype=bool,
+        """Set up velocity and VACF arrays before the analysis loop begins"""
+        # 2D array of frames x particles
+        self.results.vacf_by_particle = np.zeros(
+            (self.n_frames, self.n_particles)
         )
 
+        # 3D array of frames x particles x dimensionality
+        self._velocity_array = np.zeros(
+            (self.n_frames, self.n_particles, self.dim_fac)
+        )
+        # self.results.timeseries not set here
+
+    def _parse_dim_type(self):
+        r"""Sets up the desired dimensionality of the VACF."""
+        keys = {
+            "x": [0],
+            "y": [1],
+            "z": [2],
+            "xy": [0, 1],
+            "xz": [0, 2],
+            "yz": [1, 2],
+            "xyz": [0, 1, 2],
+        }
+
+        self.dim_type = self.dim_type.lower()
+
+        try:
+            self._dim = keys[self.dim_type]
+        except KeyError:
+            raise ValueError(
+                "invalid dim_type: {} specified, please specify one of xyz, "
+                "xy, xz, yz, x, y, z".format(self.dim_type)
+            )
+
+        self.dim_fac = len(self._dim)
+
     def _single_frame(self):
-        """Calculate data from a single frame of trajectory"""
+        """Constructs array of velocities for VACF calculation."""
         # This runs once for each frame of the trajectory
-        # It can contain the main analysis method, or just collect data
-        # so that analysis can be done over the aggregate data
-        # in _conclude.
 
         # The trajectory positions update automatically
-        negative = self.atomgroup.positions < 0
         # You can access the frame number using self._frame_index
-        self.results.is_negative[self._frame_index] = negative.any(axis=1)
 
+        # trajectory must have velocity information
+        if not self._ts.has_velocities:
+            raise NoDataError(
+                "VACF computation requires velocities " "in the trajectory"
+            )
+
+        # set shape of velocity array
+        self._velocity_array[self._frame_index] = self.atomgroup.velocities[
+            :, self._dim
+        ]
+
+    # Results will be in units of (angstroms / ps)^2
     def _conclude(self):
         """Calculate the final results of the analysis"""
         # This is an optional method that runs after
         # _single_frame loops over the trajectory.
         # It is useful for calculating the final results
         # of the analysis.
-        # For example, below we determine the
-        # which atoms always have negative coordinates.
-        self.results.always_negative = self.results.is_negative.all(axis=0)
-        always_negative_atoms = self.atomgroup[self.results.always_negative]
-        self.results.always_negative_atoms = always_negative_atoms
-        self.results.always_negative_atom_names = always_negative_atoms.names
+        if self.fft:
+            self._conclude_fft()
+        else:
+            self._conclude_simple()
 
-        # results don't have to be arrays -- they can be any value, e.g. floats
-        self.results.n_negative_atoms = self.results.is_negative.sum(axis=1)
-        self.results.mean_negative_atoms = self.results.n_negative_atoms.mean()
+    def _conclude_fft(self):  # with FFT, np.float64 bit prescision required.
+        r"""Calculates the VACF via the FCA fast correlation algorithm."""
+        try:
+            import tidynamics
+        except ImportError:
+            raise ImportError(
+                """ERROR --- tidynamics was not found!
+
+                tidynamics is required to compute an FFT based VACF (default)
+
+                try installing it using pip eg:
+
+                    pip install tidynamics
+
+                or set fft=False"""
+            )
+
+        velocities = self._velocity_array.astype(np.float64)
+        for n in range(self.n_particles):
+            self.results.vacf_by_particle[:, n] = tidynamics.acf(
+                velocities[:, n, :]
+            )
+        self.results.timeseries = self.results.vacf_by_particle.mean(axis=1)
+
+    def _conclude_simple(self):
+        r"""Calculates the VACF via the simple "windowed" algorithm."""
+        # total frames in trajectory, use N for readability
+        N = self.n_frames
+
+        # improve precision with np.float64
+        velocities = self._velocity_array.astype(np.float64)
+
+        # iterate through all possible lagtimes up to N
+        for lag in range(N):
+            # get product of velocities shifted by "lag" frames
+            veloc = velocities[: N - lag, :, :] * velocities[lag:, :, :]
+
+            # dot product of x(, y, z) velocities per particle
+            sum_veloc = np.sum(veloc, axis=-1)
+
+            # average over # frames
+            # update VACF by particle array
+            self.results.vacf_by_particle[lag, :] = np.mean(sum_veloc, axis=0)
+        # average over # particles and update results array
+        self.results.timeseries = self.results.vacf_by_particle.mean(axis=1)