+++ transport_analysis/tests/utils.py
@@ -10,7 +10,7 @@ def make_Universe(
     size: Tuple[int, int, int] = (125, 25, 5),
     n_frames: int = 0,
     velocities: bool = False,
-    forces: bool = False
+    forces: bool = False,
 ) -> mda.Universe:
     """Make a dummy reference Universe
 
@@ -49,10 +49,10 @@ def make_Universe(
         n_atoms=n_atoms,
         n_residues=n_residues,
         n_segments=n_segments,
-        atom_resindex=np.repeat(
-            np.arange(n_residues), n_atoms // n_residues),
+        atom_resindex=np.repeat(np.arange(n_residues), n_atoms // n_residues),
         residue_segindex=np.repeat(
-            np.arange(n_segments), n_residues // n_segments),
+            np.arange(n_segments), n_residues // n_segments
+        ),
         # trajectory things
         trajectory=trajectory,
         velocities=velocities,